Updating my files

+++ server/routers/snippetRouter.js
@@ -4,9 +4,7 @@ const auth = require("../middleware/auth");
 
 router.get("/", auth, async (req, res) => {
   try {
-    console.log(req.user);
-
-    const snippets = await Snippet.find();
+    const snippets = await Snippet.find({ user: req.user });
     res.json(snippets);
   } catch (err) {
     res.status(500).send();
@@ -14,7 +12,7 @@ router.get("/", auth, async (req, res) => {
 });
 
 // posting throws a 500 error that needs to be handled, error was simply no description or code was given, DUH!
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
   try {
     const { title, description, code } = req.body;
 
@@ -30,7 +28,7 @@ router.post("/", async (req, res) => {
       title,
       description,
       code,
-      // user: req.user,
+      user: req.user,
     });
 
     const savedSnippet = await newSnippet.save();
@@ -41,7 +39,7 @@ router.post("/", async (req, res) => {
   }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   try {
     const { title, description, code } = req.body;
     const snippetId = req.params.id;
@@ -66,8 +64,8 @@ router.put("/:id", async (req, res) => {
           "No snippet with this ID was found. Please contact the developer.",
       });
 
-    // if (originalSnippet.user.toString() !== req.user)
-    //   return res.status(401).json({ errorMessage: "Unauthorized." });
+    if (originalSnippet.user.toString() !== req.user)
+      return res.status(401).json({ errorMessage: "Unauthorized." });
 
     originalSnippet.title = title;
     originalSnippet.description = description;
@@ -81,7 +79,7 @@ router.put("/:id", async (req, res) => {
   }
 });
 
-router.delete("/:id",  async (req, res) => {
+router.delete("/:id",  auth, async (req, res) => {
   try {
     const snippetId = req.params.id;
 
@@ -99,8 +97,9 @@ router.delete("/:id",  async (req, res) => {
           "No snippet with this ID was found. Please contact the developer.",
       });
 
-    // if (existingSnippet.user.toString() !== req.user)
-    //   return res.status(401).json({ errorMessage: "Unauthorized." });
+    // if not the user's snippet then they cannot delete it
+    if (existingSnippet.user.toString() !== req.user)
+      return res.status(401).json({ errorMessage: "Unauthorized." });
 
     await existingSnippet.delete();
